+++ convert/intel/run_intel.py
@@ -1,28 +1,22 @@
-import sys, subprocess, argparse
+import sys, argparse, sys
 # Append to API
 from pathlib import Path
 sys.path.append(str(Path(__file__).resolve().parents[2]/"common"))
 from logger import config_logger
+from utils import cmd
 
-def cmd(command):
-    process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT,shell=True)
-    for line in iter(process.stdout.readline,b''):
-        line = line.rstrip().decode('ascii',errors='ignore')
-
-        if line.isspace(): 
-            continue
-        else:
-            print(line)
+CONTAINER_NAME = "intel-convert"
 
 def main(args):
+    SPLIT_ACTION = False
     command = "./convert/intel/run.sh"
-    cmd(command)
-    command = "docker exec -it intel-convert \
+    cmd(command, SPLIT_ACTION)
+    command = "docker exec -it {} \
                 bash -c 'pip install colorlog && \
-                python3 ./convert/intel/convert_intel.py -c {}' ".format(args.config)
-    cmd(command)
-    command = "docker stop intel-convert"
-    cmd(command)
+                python3 ./convert/intel/convert_intel.py -c {}' ".format(CONTAINER_NAME, args.config)
+    cmd(command, SPLIT_ACTION)
+    command = "docker stop {}".format(CONTAINER_NAME)
+    cmd(command, SPLIT_ACTION)
 
 if __name__ == '__main__':
     config_logger('./convert.log', 'a', "info")