+++ convert/xilinx/run_xilinx.py
@@ -3,28 +3,21 @@ import sys, argparse, subprocess
 from pathlib import Path
 sys.path.append(str(Path(__file__).resolve().parents[2]/"common"))
 from logger import config_logger
-from utils import read_json 
+from utils import cmd
 
-def cmd(command):
-    process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT,shell=True)
-    for line in iter(process.stdout.readline,b''):
-        line = line.rstrip().decode('ascii',errors='ignore')
-
-        if line.isspace(): 
-            continue
-        else:
-            print(line)
+CONTAINER_NAME = "vitis-ai-ivit-t"
 
 def main(args):
+    SPLIT_ACTION = False
     command = "./convert/xilinx/Vitis-AI/vitis-ai-start.sh c"
-    cmd(command)
-    command = "docker exec -it vitis-ai-ivit-t \
+    cmd(command, SPLIT_ACTION)
+    command = "docker exec -it {} \
                 bash -c 'source /opt/vitis_ai/conda/bin/activate vitis-ai-tensorflow && \
                 pip install colorlog && \
-                python3 ./convert/xilinx/convert_xilinx.py -c {}' ".format(args.config)
-    cmd(command)
-    command = "docker stop vitis-ai-ivit-t && docker rm vitis-ai-ivit-t"
-    cmd(command)
+                python3 ./convert/xilinx/convert_xilinx.py -c {}' ".format(CONTAINER_NAME, args.config)
+    cmd(command, SPLIT_ACTION)
+    command = "docker stop {} && docker rm {}".format(CONTAINER_NAME, CONTAINER_NAME)
+    cmd(command, SPLIT_ACTION)
 
 if __name__ == '__main__':
     config_logger('./convert.log', 'a', "info")