+++ docker/init_env.sh
@@ -12,17 +12,28 @@ function print_magic(){
 	fi
 	echo ""
 }
-# ---------------------------------------------------------
-# Set the default value of the getopts variable 
-platform="0,1,2"
-platform_list=('nvidia' 'intel' 'xilinx')
+
+function google_download(){
+	wget --load-cookies /tmp/cookies.txt \
+	"https://docs.google.com/uc?export=download&confirm=$(wget \
+	--quiet --save-cookies /tmp/cookies.txt --keep-session-cookies \
+	--no-check-certificate 'https://docs.google.com/uc?export=download&id='$1'' \
+	-O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/1n/p')&id=$1" \
+	-O "$2" && rm -rf /tmp/cookies.txt
+}
 
 # ---------------------------------------------------------
-# color ANIS
-RED='\033[0;31m'
-BLUE='\033[0;34m'
+# Set the default value of the getopts variable 
+platform="0,1,2,3"
+platform_list=('nvidia' 'intel' 'xilinx', 'hailo')
+
+# ---------------------------------------------------------
+# Color ANIS
+RED='\033[1;31m'
+BLUE='\033[1;34m'
+GREEN='\033[1;32m'
 YELLOW='\033[1;33m'
-GREEN='\033[0;32m'
+CYAN='\033[1;36m'
 NC='\033[0m';
 
 # ---------------------------------------------------------
@@ -33,7 +44,7 @@ function help(){
 	echo
 	echo "Syntax: scriptTemplate [-p|m|h]"
 	echo "options:"
-	echo "p		Select any platform.(item: [0:nvidia, 1:intel, 2:xilinx]) ex: -p '1,2' "
+	echo "p		Select any platform.(item: [0:nvidia, 1:intel, 2:xilinx, 3:hailo]) ex: -p '1,2' "
     echo "m		Print information with magic"
 	echo "h		help."
 	echo "-----------------------------------------------------------------------"
@@ -119,18 +130,18 @@ for i in ${arr_index[@]}
 do
     if [[ ${i} == *"0"* ]]; then
 		echo -e "${GREEN}"
-        echo "-----Build image of nvidia-----"
+        echo "-----Building image of nvidia-----"
 		echo -e "${NC}"
     fi
     if [[ ${i} == *"1"* ]]; then
 		echo -e "${BLUE}"
-        echo "-----Build image of intel-----"
+        echo "-----Building image of intel-----"
 		echo -e "${NC}"
 		docker build -t intel-convert -f ./convert/intel/intel.Dockerfile ./convert --no-cache
     fi
     if [[ ${i} == *"2"* ]]; then
 		echo -e "${RED}"
-        echo "-----Build image of xilinx-----"
+        echo "-----Building image of xilinx-----"
 		echo -e "${NC}"
 		
         cd ./convert/xilinx
@@ -141,16 +152,30 @@ do
 		echo "-----Download conver folder of xilinx-----"
 		FILEID="1yzYhz6T2u2GNCoqVjRcwaHDJ1_QoTBQk"
 		STORREFILE="vitis-ai-utility.zip"
-		wget --load-cookies /tmp/cookies.txt \
-			"https://docs.google.com/uc?export=download&confirm=$(wget \
-			--quiet --save-cookies /tmp/cookies.txt --keep-session-cookies \
-			--no-check-certificate 'https://docs.google.com/uc?export=download&id='${FILEID}'' \
-			-O- | sed -rn 's/.*confirm=([0-9A-Za-z_]+).*/1n/p')&id=${FILEID}" \
-			-O "$STORREFILE" && rm -rf /tmp/cookies.txt
+		google_download $FILEID $STORREFILE
 		unzip $STORREFILE
 		rm $STORREFILE
 		cp ./vitis-ai-utility/vitis-ai-start.sh ./vitis-ai-start.sh
 		cp ./vitis-ai-utility/docker_run.sh ./docker_run.sh
+    fi
+    if [[ ${i} == *"3"* ]]; then
+		echo -e "${CYAN}"
+        echo "-----Building image of hailo-----"
+		echo -e "${NC}"
+
+        cd ./convert/hailo
+		FILEID="1IFoof3TjeN2o7yBSZDhYR_GsMy_mC53V"
+		STORREFILE="hailo_sw_suite_2023-01.zip"
+		google_download $FILEID $STORREFILE
+		unzip $STORREFILE
+		rm $STORREFILE
+		# Download hailo docker/convert package
+		echo "-----Download conver folder of hailo-----"
+		FILEID="1UvoBn8eEP91goi9-wg3bS90vFRuaUgIv"
+		STORREFILE="pytorch-YOLO.zip"
+		google_download $FILEID $STORREFILE
+		unzip $STORREFILE
+		rm $STORREFILE
     fi
 done